svg view works!

+++ src/cs5004/animator/model/Animation.java
@@ -63,6 +63,6 @@ public interface Animation {
 
 
   //todo: add jdoc and readme
-  String generateXML(int msecsPtick);
+  String generateXML(Long msecsPtick);
 
 }
